fix(go/gorilla): split insecure cookie rule
