+++ tests/go/gorilla/cookie_missing_http_only/testdata/main.go
@@ -1,3 +1,5 @@
+// Use bearer:expected go_gorilla_cookie_missing_http_only to flag expected findings
+
 package session
 
 import (
@@ -11,19 +13,21 @@ var store = sessions.NewCookieStore([]byte(config.Cfg.Sessionkey))
 
 func SetSession(w http.ResponseWriter, r *http.Request, data map[string]string) {
 	session, _ := store.Get(r, "session")
+	// bearer:expected go_gorilla_cookie_missing_http_only
 	session.Options = &sessions.Options{
 		Path:     "/",
 		MaxAge:   3600,
-		HttpOnly: true,
+		HttpOnly: false,
 		Secure:   true,
 	}
 }
 
 func DeleteSession(w http.ResponseWriter, r *http.Request) {
 	session, _ := store.Get(r, "session")
+	// bearer:expected go_gorilla_cookie_missing_http_only
 	session.Options = &sessions.Options{
 		MaxAge:   -1,
-		HttpOnly: true,
+		HttpOnly: false,
 		Secure:   true,
 	}
